Implement CI based on hardware KVM on Travis
Runs at Pull Request time.

Will be scheduled to run as a cronjob.

Always pulls the latest of the upstream branches, specific
commits receive no additional testing.

Does not test pulp2-nightly-pulp3-source-centos7 yet;
may be feasible and fast enough if we try adjusting the RAM
and creating a swapfile.